+++ authlete/dto/userinfo_request.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,7 +22,10 @@ class UserInfoRequest(Jsonable):
     def __init__(self, nameAndValues=None):
         nameAndTypes = {
             'token':             str,
-            'clientCertificate': str
+            'clientCertificate': str,
+            'dpop':              str,
+            'htm':               str,
+            'htu':               str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)