+++ CHANGES.md
@@ -88,6 +88,11 @@ CHANGES
     * Added `certificateThumbprint`.
     * Added `dpopKeyThumbprint`.
 
+- `UserInfoRequest` class
+    * Added `dpop`.
+    * Added `htm`.
+    * Added `htu`.
+
 
 1.0.2 (2019-08-09)
 ------------------